requirements.txt
Django==3.0.11
djangorestframework==3.12.2
psycopg2-binary>=2.8
PyJWT==1.7.1
requests==2.22.0
django-filter==2.4.0